SOlucion de error al intetar hacer toogle en el limiete de el viewmodel.

+++ Models/Sala.cs
@@ -17,7 +17,7 @@ namespace Battleship_HTTP.Models
         public string? IdJugador2 { set; get; }
         public string? NombreJugador2 { set; get; }
         public bool ListoJugador2 { get; set; } = false;
-        public byte JugadoresListos { set; get; }
+        public int JugadoresListos { set; get; }
         public bool Publica { get; set; }
 
         public bool Llena => IdJugador1 != null && IdJugador2 != null;
